slightly better syntax

+++ chicken/Assets/Scripts/PlayerControl.cs
@@ -24,8 +24,8 @@ public class PlayerControl : MonoBehaviour
     public bool ARCollected;
 
     [Header("Player Stats")]
-    public int MaxHealth = 5;
-    public int CurrentHealth = 5;
+    public float MaxHealth = 5;
+    public float CurrentHealth = 5;
     public int HealVal = 1;
     public bool inAir = true;
     public float groundDetecDist = 1.1f;
